Expand automatic harvest state logic to cover more scenarios.
If a block has an age property but is not a crop, the age will be set to
the max value by default.

If a block has a berries property that is a boolean, like glow berries, it
will be set to true by default.

If a block is multifaced and bonemealable the down face will be enabled by
default.